split Ruby into new structure
