+++ packages/quicktype-core/src/language/Ruby/utils.ts
@@ -1,0 +1,64 @@
+import unicode from "unicode-properties";
+
+import * as keywords from "./constants";
+
+import {
+    legalizeCharacters,
+    splitIntoWords,
+    combineWords,
+    firstUpperWordStyle,
+    allUpperWordStyle,
+    allLowerWordStyle,
+    utf32ConcatMap,
+    isPrintable,
+    escapeNonPrintableMapper,
+    intToHex,
+    isLetterOrUnderscore
+} from "../../support/Strings";
+
+export const forbiddenForObjectProperties = Array.from(new Set([...keywords.keywords, ...keywords.reservedProperties]));
+function unicodeEscape(codePoint: number): string {
+    return "\\u{" + intToHex(codePoint, 0) + "}";
+}
+
+export const stringEscape = utf32ConcatMap(escapeNonPrintableMapper(isPrintable, unicodeEscape));
+
+const isStartCharacter = isLetterOrUnderscore;
+
+function isPartCharacter(utf16Unit: number): boolean {
+    const category: string = unicode.getCategory(utf16Unit);
+    return ["Nd", "Pc", "Mn", "Mc"].indexOf(category) >= 0 || isStartCharacter(utf16Unit);
+}
+
+const legalizeName = legalizeCharacters(isPartCharacter);
+
+export function simpleNameStyle(original: string, uppercase: boolean): string {
+    if (/^[0-9]+$/.test(original)) {
+        original = original + "N";
+    }
+    const words = splitIntoWords(original);
+    return combineWords(
+        words,
+        legalizeName,
+        uppercase ? firstUpperWordStyle : allLowerWordStyle,
+        uppercase ? firstUpperWordStyle : allLowerWordStyle,
+        allUpperWordStyle,
+        allUpperWordStyle,
+        "",
+        isStartCharacter
+    );
+}
+
+export function memberNameStyle(original: string): string {
+    const words = splitIntoWords(original);
+    return combineWords(
+        words,
+        legalizeName,
+        allLowerWordStyle,
+        allLowerWordStyle,
+        allLowerWordStyle,
+        allLowerWordStyle,
+        "_",
+        isStartCharacter
+    );
+}
